Move namespace testing to separate class.
